# Conflicts:
#	index.js

+++ app/party-manager.js
@@ -104,7 +104,7 @@ class PartyManager {
             party.sendDeletionWarningMessage();
             await party.persist();
           }
-          if (party.deletionTime && now > party.deletionTime) {
+          if (party.deletionTime && now > party.deletionTime && party.deletionTime !== -1) {
             party.delete();
           }
 
@@ -316,7 +316,11 @@ class PartyManager {
   async getMember(channelId, memberId) {
     const party = this.getParty(channelId),
       channel = (await this.getChannel(channelId)).channel,
-      member = channel.guild.members.get(memberId);
+      member = await channel.guild.members.fetch(memberId)
+        .catch(err => {
+          log.error(err);
+          return undefined;
+        });
 
     if (!!member) {
       return Promise.resolve({member, ok: true});
@@ -343,9 +347,13 @@ class PartyManager {
       .includes(gymId);
   }
 
-  getChannel(channelId) {
+  async getChannel(channelId) {
     try {
-      const channel = this.client.channels.get(channelId);
+      const channel = await this.client.channels.fetch(channelId)
+        .catch(err => {
+          log.error(err);
+          return undefined;
+        });
 
       if (!channel) {
         if (this.validParty(channelId)) {
@@ -393,8 +401,17 @@ class PartyManager {
               return {error: new Error('Message does not exist'), ok: false};
             }
           } else {
-            const message = await channel.channel.messages.fetch(messageId);
-            return {message: message, ok: true};
+            const message = await channel.channel.messages.fetch(messageId)
+              .catch(err => {
+                log.error(err);
+                return undefined;
+              });
+
+            if (!!message) {
+              return {message: message, ok: true};
+            } else {
+              return {error: new Error('Could not fetch message'), ok: false};
+            }
           }
         })
         .catch(err => {
